+++ the_warroom/models.py
@@ -389,6 +389,9 @@ class Tournament(models.Model):
     def get_settings_url(self):
         return reverse('tournament-settings', kwargs={'slug': self.slug})
 
+    def get_schedule_url(self):
+        return reverse('tournament-schedule-page', kwargs={'slug': self.slug})
+
     def get_edit_url(self):
         return reverse('tournament-dynamic-update', kwargs={'slug': self.slug})
 
@@ -829,6 +832,9 @@ class Stage(models.Model):
     def get_create_round_url(self):
         return reverse('round-create', kwargs={'tournament_slug': self.tournament.slug, 'stage_slug': self.slug})
 
+    def get_schedule_url(self):
+        return reverse('stage-schedule-page', kwargs={'tournament_slug': self.tournament.slug, 'stage_slug': self.slug})
+
     def add_player(self, profile):
         """Add a profile to this stage, creating a TournamentPlayer if needed."""
         tp, _ = TournamentPlayer.objects.get_or_create(
@@ -1256,6 +1262,19 @@ class Round(models.Model):
                 'round_slug': self.slug
             })
         return reverse('round-matches-page', kwargs={
+            'tournament_slug': tournament.slug,
+            'stage_slug': self.stage.slug,
+            'round_slug': self.slug
+        })
+
+    def get_schedule_url(self):
+        tournament = self.stage.tournament
+        if not tournament.use_stages:
+            return reverse('round-schedule-simple', kwargs={
+                'tournament_slug': tournament.slug,
+                'round_slug': self.slug
+            })
+        return reverse('round-schedule-page', kwargs={
             'tournament_slug': tournament.slug,
             'stage_slug': self.stage.slug,
             'round_slug': self.slug
@@ -2042,9 +2061,9 @@ class PlayerGroup(models.Model):
     def save(self, *args, **kwargs):
         if self.video_link:
             if 'twitch.tv' in self.video_link:
-                self.video_platform = self.VideoPlatformChoices.TWITCH
+                self.video_platform = VideoPlatformChoices.TWITCH
             elif 'youtube.com' in self.video_link or 'youtu.be' in self.video_link:
-                self.video_platform = self.VideoPlatformChoices.YOUTUBE
+                self.video_platform = VideoPlatformChoices.YOUTUBE
             else:
                 self.video_platform = ''
         else: